% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_covid.R
\name{covid_simu_summary_ind}
\alias{covid_simu_summary_ind}
\title{Summarize the simulation results}
\usage{
covid_simu_summary_ind(dat_rst, par_analysis, nsmps = 3000)
}
\description{
Summarize the simulation results
}
