% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_covid.R
\name{covid_surv_ana_logrank}
\alias{covid_surv_ana_logrank}
\title{Generate analysis data set}
\usage{
covid_surv_ana_logrank(
  data_ana,
  fml = "Surv(day_obs, status_obs) ~ arm",
  mth_fix_fu = NULL
)
}
\description{
Generate analysis dataset for minimum fu days
}
