% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{desfix_kernel_density_multiple}
\alias{desfix_kernel_density_multiple}
\title{Plot multiple kernel density}
\usage{
desfix_kernel_density_multiple(
  x,
  xlab_name,
  para_notation,
  percent,
  logYN,
  from,
  to
)
}
\description{
Plot multiple kernel density
}
