% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design_rmeasure.R
\name{internal_rmeasure_dpara}
\alias{internal_rmeasure_dpara}
\title{Default design parameter for single arm}
\usage{
internal_rmeasure_dpara()
}
\description{
Default design parameter for single arm
}
