% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_biom.R
\name{internal_survbiom_dpara}
\alias{internal_survbiom_dpara}
\title{Get Study Design for PFS with Biomarker}
\usage{
internal_survbiom_dpara(
  sample_size = 600,
  ratio_by_arm = c(1/3, 1/3, 1/3),
  p_biom_by_arm = rbin(c(0.3, 0.7), c(0.3, 0.7), c(0.3, 0.7)),
  ctl_median_surv = c(10, 15),
  hr = c(0.7, 0.7),
  annual_drop = 1e-06,
  enroll_dur_mth = 18,
  info_frac = c(0.5, 0.7, 1),
  btype_primary = "asOF",
  alpha = 0.025,
  target_events = 200,
  drop_info_frac = 0.3,
  drop_event = "pfs",
  f_arm_sel = stb_surv_biom_arm_sel_rule_1,
  margin_r3 = 0.1,
  fml_surv = "Surv(day_pfs, status_pfs) ~ arm",
  ...
)
}
\description{
Get Study Design for PFS with Biomarker
}
