% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_survjoin.R
\name{internal_survjoin_dpara}
\alias{internal_survjoin_dpara}
\title{Get Study Design for Joint Primary and Secondary Endpoints}
\usage{
internal_survjoin_dpara(
  info_frac = c(0.3, 0.7, 1),
  primary = c("os", "pfs"),
  target_primary = 100,
  sample_size = 500,
  ctl_ratio = 0.5,
  annual_drop = 1e-06,
  enroll_dur_mth = 18,
  alpha = 0.025,
  btype_primary = "asOF",
  bound_second = NULL,
  ctl_median_os = 15,
  ctl_median_pfs = 9,
  hr_os = 0.7,
  hr_pfs = 1,
  rho_ctl = 0.8,
  rho_trt = NULL,
  method = c("given.pfs", "given.os"),
  ...,
  seed = NULL
)
}
\description{
Get Study Design for Joint Primary and Secondary Endpoints
}
