% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{lognormal_density_multiple}
\alias{lognormal_density_multiple}
\title{Plot multiple density for FIX}
\usage{
lognormal_density_multiple(true_m, true_cv, xlab_name, logYN)
}
\description{
Plot multiple density for FIX
}
