% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rcurrent.R
\name{rcur_eff_event_dta}
\alias{rcur_eff_event_dta}
\title{Effective number of events}
\usage{
rcur_eff_event_dta(dta, weight = 0.2)
}
\description{
Calculate effective number of events for recurrent events
}
