% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rcurrent.R
\name{rcurrent_day_eos_2}
\alias{rcurrent_day_eos_2}
\title{Generate analysis data set}
\usage{
rcurrent_day_eos_2(data_full, fu_days = 12 * 7, pt_proportion = 1)
}
\description{
Generate analysis dataset for fixed fu days
}
