% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rcurrent.R
\name{rcurrent_day_eos_adapt_2}
\alias{rcurrent_day_eos_adapt_2}
\title{Generate analysis data set}
\usage{
rcurrent_day_eos_adapt_2(
  data_full,
  n_stage1,
  n_stage2 = 0,
  target_event = 0,
  rcur_weight = 0.2,
  fix_fu = 12 * 7
)
}
\description{
Generate final analysis dataset for 1) stage 1 patients with fixed fu, 2)
stage 2 patients with fixed fu, 3) target number of events
}
