% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rcurrent.R
\name{rcurrent_day_eos_adapt_3}
\alias{rcurrent_day_eos_adapt_3}
\title{Generate analysis data set}
\usage{
rcurrent_day_eos_adapt_3(data_full, n, fix_fu = 12 * 7)
}
\description{
Generate final analysis dataset for all patients with fixed fu
}
