% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rcurrent.R
\name{rcurrent_logrank_size}
\alias{rcurrent_logrank_size}
\title{Calculate NB sample size based on survival analysis}
\usage{
rcurrent_logrank_size(
  r0,
  r1,
  fix_fu,
  power = 0.9,
  rcur_weight = 0.2,
  k = 1,
  alpha = 0.05
)
}
\arguments{
\item{r0}{hazard rate in control}

\item{r1}{hazard rate in treatment}

\item{rcur_weight}{information weight on recurrent events}

\item{fu}{average follow up time on each pt}
}
\description{
Calculate NB sample size based on logrank power and sample size calculation
}
