% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rcurrent.R
\name{rcurrent_ssr_ana_set}
\alias{rcurrent_ssr_ana_set}
\title{Rcurrent sample size re-estimation}
\usage{
rcurrent_ssr_ana_set(data, lst_design)
}
\description{
This is based on a naive NB sample size re-estimation
}
