% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rmeasure.R
\name{rmeasure_ana_bayes}
\alias{rmeasure_ana_bayes}
\title{Get posterior samples of treatment effect at each visit}
\usage{
rmeasure_ana_bayes(data_ana, n_visit = 3, formula = NULL, ...)
}
\description{
Get posterior samples of treatment effect at each visit
}
