% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_rmeasure.R
\name{rmeasure_ana_mmrm}
\alias{rmeasure_ana_mmrm}
\title{Generate analysis data set}
\usage{
rmeasure_ana_mmrm(
  data_ana,
  endpoint_visit = 1,
  formula = NULL,
  contrast = NULL
)
}
\description{
Generate analysis dataset for minimum fu days
}
