% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design_rmeasure.R
\name{rmeasure_create_ana_data}
\alias{rmeasure_create_ana_data}
\title{Generate analysis data set}
\usage{
rmeasure_create_ana_data(data_full, visit_no, fraction)
}
\description{
Generate analysis dataset for minimum fu days
}
