% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_conduct.R
\name{stb_conduct_surv_biom}
\alias{stb_conduct_surv_biom}
\title{Conduct Simulation Study for Arm Selection}
\usage{
stb_conduct_surv_biom(lst_design, n_rep = 1000, n_core = 1, seed = NULL)
}
\description{
Conduct Simulation Study for Arm Selection
}
