% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_conduct.R
\name{stb_conduct_surv_join}
\alias{stb_conduct_surv_join}
\title{Conduct Simulation Study for Joint Primary and Secondary Endpoints}
\usage{
stb_conduct_surv_join(lst_design, n_rep = 1000, n_core = 1, seed = NULL)
}
\description{
Conduct Simulation Study for Joint Primary and Secondary Endpoints
}
