% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_class_design.R
\name{stb_create_design}
\alias{stb_create_design}
\title{Create a Study Design}
\usage{
stb_create_design(
  type = c("surv_strat", "surv_or", "surv_join", "surv_biom", "bayes_1arm", "bayes_2arm",
    "rcurrent", "rmeasure", "msma_surv", "dose_fix", "covid")
)
}
\description{
Create a Study Design
}
