% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design.R
\name{stb_design_surv_join}
\alias{stb_design_surv_join}
\title{Get Study Design for Joint Primary and Secondary Endpoints}
\usage{
stb_design_surv_join(
  info_frac = c(0.3, 0.7, 1),
  primary = c("os", "pfs"),
  target_primary = 100,
  sample_size = 500,
  ctl_ratio = 0.5,
  annual_drop = 1e-06,
  enroll_dur_mth = 18,
  alpha = 0.025,
  btype_primary = "asOF",
  bound_second = NULL,
  ctl_median_os = 15,
  ctl_median_pfs = 9,
  hr_os = 0.7,
  hr_pfs = 1,
  rho_ctl = 0.8,
  rho_trt = NULL,
  ...,
  seed = NULL
)
}
\arguments{
\item{btype_primary}{type for primary bound}

\item{bound_secondary}{alpha level at each interim analysis. Default is the
naive alpha boundary.}
}
\description{
Get Study Design for Joint Primary and Secondary Endpoints
}
