% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design.R
\name{stb_design_surv_stra}
\alias{stb_design_surv_stra}
\title{Get Study Design for Stratified Survival Analysis}
\usage{
stb_design_surv_stra(
  target = 100,
  sample_size = 500,
  ctl_ratio = 0.5,
  annual_drop = 1e-06,
  enroll_dur_mth = 18,
  stra_freq = c(0.5, 0.5),
  ctl_median = c(15, 15),
  hr = c(0.7, 0.7)
)
}
\arguments{
\item{stra_freq}{frequency of each stratum}

\item{ctl_median}{median survival time for each stratum}

\item{hr}{hazard ratio for each stratum}
}
\description{
Get Study Design for Stratified Survival Analysis
}
