% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dists.R
\name{stb_dist_beta_para_by_qs}
\alias{stb_dist_beta_para_by_qs}
\title{Get beta parameters by quantiles}
\usage{
stb_dist_beta_para_by_qs(
  target_qs = c(`0.1` = 0.5, `0.5` = 0.7),
  para = c(shape1 = 1, shape2 = 1),
  lower = c(0, 0),
  method = "L-BFGS-B",
  ...
)
}
\description{
@export
}
