% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_class.R
\name{stb_dist_by_qs}
\alias{stb_dist_by_qs}
\title{Create a distribution}
\usage{
stb_dist_by_qs(
  type = c("normal", "lognormal", "beta"),
  target_qs = c(`0.5` = 0.7, `0.8` = 0.9)
)
}
\description{
Create a distribution
}
