% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dists.R
\name{stb_dist_lognormal_para_by_qs}
\alias{stb_dist_lognormal_para_by_qs}
\title{Get log-normal parameters by quantiles}
\usage{
stb_dist_lognormal_para_by_qs(target_qs = c(`0.1` = 0.5, `0.5` = 0.7), ...)
}
\description{
@export
}
