% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dst.R
\name{stb_dist_normal_para_by_qs}
\alias{stb_dist_normal_para_by_qs}
\title{Get normal parameters by quantiles}
\usage{
stb_dist_normal_para_by_qs(
  target_qs = c(`0.1` = 0.5, `0.5` = 0.7),
  para = c(mean = 1, sd = 1),
  lower = c(-Inf, 0),
  method = "L-BFGS-B",
  ...
)
}
\description{
@export
}
