% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design_rcurrent.R
\name{stb_rc_power}
\alias{stb_rc_power}
\title{Power calculation}
\usage{
stb_rc_power(n, mu_t, r0, r1, k = 1, alpha = 0.05)
}
\arguments{
\item{n}{total sample size, assuming 1:1 randomization ratio}

\item{mu_t}{mean follow-up time}

\item{r0}{mean hazard rate in control arm}

\item{r1}{mean hazard rate in treatment arm}

\item{k}{dispersion parameter in negative binomial}

\item{alpha}{type I error rate}
}
\description{
Power calculation using the formula proposed in Zhu and Lakkis (2014)
}
\examples{
   stb_rc_power(n = 100, mu_t = 2, r0 = 3, r1 = 1.9, k = 1)

}
