% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design_rcurrent.R
\name{stb_rc_simu_events}
\alias{stb_rc_simu_events}
\title{Simulate Negative Binomial with Event time}
\usage{
stb_rc_simu_events(n, mu, k = 1, n_event = 1, seed = NULL)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean of the negative binomial}

\item{k}{dispersion parameter}

\item{n_event}{number of event time}
}
\description{
Simulate negative binomial by event time
}
