% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design_rcurrent.R
\name{stb_rc_simu_events_to_nb}
\alias{stb_rc_simu_events_to_nb}
\title{Convert Recurrent Events to NB}
\usage{
stb_rc_simu_events_to_nb(dat_rc, t = 1)
}
\arguments{
\item{dat_rc}{recurrent events data frame}

\item{t}{follow up time}
}
\description{
Convert Recurrent Events to NB
}
