% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_rcurrent.R
\name{stb_rc_simu_pgamma}
\alias{stb_rc_simu_pgamma}
\title{Simulate Negative Binomial}
\usage{
stb_rc_simu_pgamma(n, mu, k, t = 1, seed = NULL)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean of the negative binomial}

\item{k}{dispersion parameter}

\item{t}{follow up time. Default value 1.}
}
\description{
Simulate negative binomial by poisson-gamma
}
