% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_onco.R
\name{stb_simu_enroll}
\alias{stb_simu_enroll}
\title{Simulate Enrollment Time}
\usage{
stb_simu_enroll(
  ntot,
  enroll_duration,
  min_fu = 6,
  date_bos = NULL,
  mth_to_days = 30.4
)
}
\arguments{
\item{ntot}{Total number of patients}

\item{enroll_duration}{Duration of enrollment in months}

\item{min_fu}{Minimum follow up months}

\item{date_bos}{Begin of study date}

\item{mth_to_days}{Days in each month}
}
\description{
Simulate Enrollment Time
}
