% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_or.R
\name{stb_simu_event}
\alias{stb_simu_event}
\title{Simulate progression, death, and censoring}
\usage{
stb_simu_event(dta_or, median_prog, median_dth, annual_drop, shift = "day_or")
}
\arguments{
\item{median_prog}{median month for progression}

\item{median_dth}{median month for death}

\item{annual_drop}{annual drop rate}

\item{ntot}{total number of patients}
}
\description{
Simulate progression, death, and censoring
}
