% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_survor.R
\name{stb_simu_or}
\alias{stb_simu_or}
\title{Simulate Overall Response}
\usage{
stb_simu_or(dta_enroll, p_or)
}
\arguments{
\item{p_or}{Probability of overall response}
}
\description{
Simulate Overall Response
}
