% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_onco.R
\name{stb_simu_trial_single}
\alias{stb_simu_trial_single}
\title{Simulate a trial}
\usage{
stb_simu_trial_single(
  ntot,
  enroll_duration,
  date_bos = as.Date("2022-1-1"),
  min_fu = 6,
  p_or = 0.4,
  or_median_prog = 6,
  or_median_dth = 10,
  nor_median_prog = 4,
  nor_median_dth = 6,
  annual_drop = 0.05
)
}
\arguments{
\item{ntot}{Total number of patients}

\item{enroll_duration}{Duration of enrollment in months}

\item{date_bos}{Begin of study date}

\item{min_fu}{Minimum follow up months}

\item{p_or}{Probability of overall response}

\item{or_median_prog}{median month for progression for OR patients}

\item{or_median_dth}{median month for death for OR patients}

\item{nor_median_prog}{median month for progression for patients without OR}

\item{nor_median_dth}{median month for death for OR patients without OR}

\item{annual_drop}{annual drop rate}

\item{mth_to_days}{Days in each month}

\item{dta_enroll}{Data frame of enrollment}
}
\description{
Simulate a trial
}
