% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_onco.R
\name{stb_simu_trials}
\alias{stb_simu_trials}
\title{Simulate trials}
\usage{
stb_simu_trials(
  nreps = 100,
  f_single = stb_simu_trial_single,
  ...,
  n_cores = 5
)
}
\arguments{
\item{...}{Parameter for f_single}
}
\description{
Simulate trials
}
