% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_biom.R
\name{stb_surv_biom_arm_sel_rule_1}
\alias{stb_surv_biom_arm_sel_rule_1}
\title{Arm Selection Rule I}
\usage{
stb_surv_biom_arm_sel_rule_1(data, ...)
}
\description{
Arm selection by interim biomarker. Select the treatment arm with the largest
biomarker mean.
}
