% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_survbiom.R
\name{stb_surv_biom_arm_sel_rule_2}
\alias{stb_surv_biom_arm_sel_rule_2}
\title{Arm Selection Rule II}
\usage{
stb_surv_biom_arm_sel_rule_2(data, fml_surv)
}
\description{
Arm selection by pfs. Select the treatment arm with the best
hazard rate.
}
