% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_survbiom.R
\name{stb_surv_biom_arm_sel_rule_3}
\alias{stb_surv_biom_arm_sel_rule_3}
\title{Arm Selection Rule I}
\usage{
stb_surv_biom_arm_sel_rule_3(data, margin = 0.1, ...)
}
\description{
Arm selection by interim biomarker. Select arm 2 only when its biom rate is
xx% better than arm 1
}
