% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design_survbiom.R
\name{stb_surv_biom_arm_simu}
\alias{stb_surv_biom_arm_simu}
\title{Simulate an arm}
\usage{
stb_surv_biom_arm_simu(
  n,
  p_biomarker,
  median_mth,
  enroll_dur_mth,
  annual_drop = 0,
  date_bos = as.Date("2022-1-1"),
  seed = NULL,
  ...
)
}
\arguments{
\item{p_biomarker}{biomarker rates}

\item{median_mth}{median survival for responders and non-responsders}
}
\description{
Simulate an arm
}
