% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_biom.R
\name{stb_surv_biom_theory}
\alias{stb_surv_biom_theory}
\title{Theoretical derivation}
\usage{
stb_surv_biom_theory(
  drop_info_frac = 0.3,
  info_frac = c(0.4, 1),
  n_arm = 3,
  btype_primary = "asOF",
  alpha = 0.025,
  n_large = 1e+06,
  seed = NULL,
  ...
)
}
\description{
Theoretical derivation
}
