% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_biom.R
\name{stb_surv_biom_theory_pcr}
\alias{stb_surv_biom_theory_pcr}
\title{Theoretical derivation II}
\usage{
stb_surv_biom_theory_pcr(
  n_1,
  p_biom_1,
  rho_1,
  n_2 = n_1,
  p_biom_2 = p_biom_1,
  rho_2 = rho_1,
  margin = 0.1,
  alpha = 0.025,
  n_large = 1e+06,
  seed = NULL,
  ...
)
}
\arguments{
\item{margin}{margin to select arm 2}

\item{alpha}{alpha level at the primary analysis}

\item{n}{sample size each arm at arm-selection look}

\item{p_biom}{true pcR rate}

\item{rho}{correlation between pCR statistic and log-rank test statistic}
}
\description{
Theoretical derivation for selecting dose arm based on pCR
}
