% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_biom.R
\name{stb_surv_biom_trial_interim}
\alias{stb_surv_biom_trial_interim}
\title{Simulate a trial with interim}
\usage{
stb_surv_biom_trial_interim(lst_design, seed = NULL)
}
\description{
Simulate a trial with interim
}
