% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_survjoin.R
\name{stb_surv_join_par}
\alias{stb_surv_join_par}
\title{Calculate hazard and coefficient for simulation}
\usage{
stb_surv_join_par(median_os, median_pfs, rho)
}
\description{
Calculate hazard and coefficient for simulation
}
