% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_join_pfs_os.R
\name{stb_surv_join_simu}
\alias{stb_surv_join_simu}
\title{Simulate OS and PFS by Copula and Exponential}
\usage{
stb_surv_join_simu(n, hazard_os, hazard_prog, rho, rnd_cdf = NULL, seed = NULL)
}
\arguments{
\item{n}{sample size}

\item{hazard_os}{overall survival hazard}

\item{hazard_prog}{progression hazard}

\item{rho}{correlation between OS and progression in gaussian copula}

\item{rnd_cdf}{exisiting samples}
}
\description{
Simulate OS and PFS by Copula and Exponential
}
