% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_survjoin.R
\name{stb_surv_join_simu_events}
\alias{stb_surv_join_simu_events}
\title{Simulate OS and PFS by Copula and Exponential}
\usage{
stb_surv_join_simu_events(n, par_pfs_os, seed = NULL)
}
\arguments{
\item{n}{sample size}

\item{hazard_os}{overall survival hazard}

\item{hazard_prog}{progression hazard}

\item{hazard_pfs}{progression free survival hazard}
}
\description{
Simulate OS and PFS by Copula and Exponential
}
