% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_pfs_os.R
\name{stb_surv_join_simu_given_os}
\alias{stb_surv_join_simu_given_os}
\title{Simulate OS and PFS by Copula and Exponential}
\usage{
stb_surv_join_simu_given_os(
  hazard_prog,
  hazard_os,
  hazard_pfs,
  rho,
  rnd_cdf = NULL,
  nsim,
  seed = NULL,
  ...
)
}
\arguments{
\item{hazard_prog}{progression hazard}

\item{hazard_os}{overall survival hazard}

\item{rho}{correlation between OS and progression in gaussian copula}

\item{rnd_cdf}{exisiting samples}

\item{n}{sample size}
}
\description{
Simulate OS and PFS by Copula and Exponential
}
