% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_pfs_os.R
\name{stb_surv_join_trial_hiertest}
\alias{stb_surv_join_trial_hiertest}
\title{Simulate a trial with interim}
\usage{
stb_surv_join_trial_hiertest(interim_rst, pval_bounds)
}
\arguments{
\item{pval_bounds}{one-sided p-value boundaries. First row for primary and
second row for secondary}
}
\description{
Simulate a trial with interim
}
