% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_pfs_os.R
\name{stb_surv_join_trial_interim}
\alias{stb_surv_join_trial_interim}
\title{Simulate a trial with interim}
\usage{
stb_surv_join_trial_interim(
  ...,
  total_events,
  info_frac,
  primary = "os",
  seed = NULL
)
}
\arguments{
\item{annual_drop}{annual drop rate}
}
\description{
Simulate a trial with interim
}
