% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_pfs_os.R
\name{stb_surv_join_trial_simu}
\alias{stb_surv_join_trial_simu}
\title{Simulate a trial}
\usage{
stb_surv_join_trial_simu(ntrt, par_trt, nctl, par_ctl, ...)
}
\arguments{
\item{annual_drop}{annual drop rate}
}
\description{
Simulate a trial
}
