% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_stra.R
\name{stb_surv_stra_stra_simu}
\alias{stb_surv_stra_stra_simu}
\title{Simulate an arm}
\usage{
stb_surv_stra_stra_simu(
  n,
  median_mth,
  enroll_dur_mth,
  annual_drop = 0.05,
  mth_to_days = 30.4,
  date_bos = as.Date("2022-1-1"),
  ...
)
}
\arguments{
\item{annual_drop}{annual drop rate}
}
\description{
Simulate an arm
}
