% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_stratify.R
\name{stb_surv_strat_trial_simu}
\alias{stb_surv_strat_trial_simu}
\title{Simulate a trial}
\usage{
stb_surv_strat_trial_simu(lst_design, seed = NULL, ...)
}
\arguments{
\item{annual_drop}{annual drop rate}
}
\description{
Simulate a trial
}
