% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_survival.R
\name{stb_tl_cumu_median}
\alias{stb_tl_cumu_median}
\title{Convert cumulative incidence to median survival}
\usage{
stb_tl_cumu_median(cumu_inc, t = 6)
}
\description{
Convert cumulative incidence to median survival
}
