% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tools_surv.R
\name{stb_tl_hazard}
\alias{stb_tl_hazard}
\title{Convert median survival or annual_drop to hazard}
\usage{
stb_tl_hazard(median_surv = NULL, annual_drop = NULL)
}
\description{
Convert median survival or annual_drop to hazard
}
