% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_survival.R
\name{stb_tl_interim_data}
\alias{stb_tl_interim_data}
\title{Get Interim Analysis Data}
\usage{
stb_tl_interim_data(
  data,
  info_frac,
  total = NULL,
  event = c("os", "pfs", "obs", "enroll"),
  arms = NULL,
  offset_days = 0
)
}
\arguments{
\item{info_frac}{information fraction}

\item{total}{total number of events or samples}

\item{event}{obs: ignore pfs/os, just an event}

\item{offset_days}{offset days after the target event for creating the
interim data}
}
\description{
This is an improved version of stb_tl_interim_data that allows to create an
interim snapshot based on number of events or samples from any specific arms
}
