% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tools_survival.R
\name{stb_tl_interim_data_2arm}
\alias{stb_tl_interim_data_2arm}
\title{Get Interim Analysis Data}
\usage{
stb_tl_interim_data_2arm(data, total_events, info_frac, event = "os")
}
\arguments{
\item{total_events}{total number of events}

\item{info_frac}{information fraction}
}
\description{
Get Interim Analysis Data
}
