% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_nb.R
\name{stb_tl_rc_pooled}
\alias{stb_tl_rc_pooled}
\title{Estimate event rate based on pooled data}
\usage{
stb_tl_rc_pooled(dat, hr = 1, fml = "y ~ offset(log(day_onstudy))")
}
\description{
Estimate event rate based on pooled data
}
