% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_survival.R
\name{stb_tl_rexp}
\alias{stb_tl_rexp}
\title{Simulate time to events in days}
\usage{
stb_tl_rexp(
  ntot,
  median_mth = NULL,
  hazard = NULL,
  annual_drop = NULL,
  mth_to_days = 30.4,
  take_floor = FALSE,
  ...
)
}
\arguments{
\item{ntot}{total number of patients}

\item{median_mth}{median survival in months}

\item{hazard}{hazard for event}

\item{annual_drop}{annual drop rate}
}
\description{
Simulate time to events in days
}
