% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_survival.R
\name{stb_tl_surv_simu_pwexp}
\alias{stb_tl_surv_simu_pwexp}
\title{Simulate from piecewise constant exponential}
\usage{
stb_tl_surv_simu_pwexp(hazards, offset = 0)
}
\description{
Simulate from piecewise constant exponential
}
\examples{
rd_sim_pwexp(hazards = c("0.5" = 0.045, "Inf" = 0.025), offset = 0)

}
