% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_design_survstra.R
\name{strasurv_simu_arm}
\alias{strasurv_simu_arm}
\title{Simulate an arm}
\usage{
strasurv_simu_arm(
  n,
  median_mth,
  enroll_dur_mth,
  annual_drop = 0.05,
  mth_to_days = 30.4,
  date_bos = as.Date("2022-1-1"),
  ...
)
}
\description{
Simulate an arm
}
