% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_surv_bor.R
\name{surv_bor_ana_post}
\alias{surv_bor_ana_post}
\title{Borrow from historical information}
\usage{
surv_bor_ana_post(
  loghr_mean,
  loghr_sd,
  prior_loghr_mean,
  prior_loghr_sd,
  a = 0
)
}
\description{
Borrow from historical information
}
