% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_surv_join.R
\name{survjoin_simu_summary}
\alias{survjoin_simu_summary}
\title{Summarize simulation results}
\usage{
survjoin_simu_summary(data_interim, primary_secondary)
}
\description{
Summarize simulation results
}
