% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_survor.R
\name{survor_gen_data}
\alias{survor_gen_data}
\title{Simulate a trial}
\usage{
survor_gen_data(
  sample_size = 100,
  enroll_dur_mth = 18,
  date_bos = as.Date("2022-1-1"),
  min_fu = 6,
  p_or = 0.4,
  or_median_prog = 6,
  or_median_dth = 10,
  nor_median_prog = 4,
  nor_median_dth = 6,
  annual_drop = 1e-05,
  ...
)
}
\arguments{
\item{sample_size}{Total number of patients}

\item{enroll_dur_mth}{Duration of enrollment in months}

\item{date_bos}{Begin of study date}

\item{min_fu}{Minimum follow up months}

\item{p_or}{Probability of overall response}

\item{or_median_prog}{median month for progression for OR patients}

\item{or_median_dth}{median month for death for OR patients}

\item{nor_median_prog}{median month for progression for patients without OR}

\item{nor_median_dth}{median month for death for OR patients without OR}

\item{annual_drop}{annual drop rate}

\item{mth_to_days}{Days in each month}
}
\description{
Simulate a trial
}
