% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_rcurrent.R
\name{stb_rc_plt_fu}
\alias{stb_rc_plt_fu}
\title{Plot recurrent events}
\usage{
stb_rc_plt_fu(dat, t = 0, subset_n = 0, seed = NULL)
}
\arguments{
\item{t}{Cut time point for the analysis}

\item{subset_n}{Number of patients to be included in the plot}
}
\description{
Create clinical timeline plot for each patient
}
