% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_join_pfs_os.R
\name{stb_surv_join_par}
\alias{stb_surv_join_par}
\title{Calculate hazard and coefficient for simulation}
\usage{
stb_surv_join_par(
  median_os,
  median_pfs,
  rho,
  nsim = 20000,
  interval = c(0, 10),
  ...,
  verbose = 0,
  seed = 1000
)
}
\arguments{
\item{rho}{correlation in copula model}

\item{nsim}{number of simulated subjects for numerical calculation}
}
\description{
Calculate hazard using copula correlation
}
