% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_nb.R
\name{stb_tl_rc_size}
\alias{stb_tl_rc_size}
\title{Sample size calculation}
\usage{
stb_tl_rc_size(power, mu_t, r0, r1, k = 1, alpha = 0.05)
}
\description{
Sample size calculation using the formula proposed in Zhu and Lakkis (2014)
}
