% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tools_surv.R
\name{stb_tl_simu_enroll}
\alias{stb_tl_simu_enroll}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll(
  ntot,
  enroll_dur_mth,
  min_fu = NULL,
  date_bos = NULL,
  mth_to_days = 30.4,
  ...
)
}
\description{
Simulate Enrollment Time
}
