% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tools_surv.R
\name{stb_tl_simu_enroll}
\alias{stb_tl_simu_enroll}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll(
  ntot,
  enroll_dur_mth,
  min_fu = NULL,
  date_bos = NULL,
  mth_to_days = 30.4,
  ...
)
}
\arguments{
\item{ntot}{Total number of patients}

\item{enroll_dur_mth}{enroll duration in months}

\item{min_fu}{minimum follow up in months}

\item{date_bos}{date begin of study}

\item{mth_to_days}{Days in each month}
}
\description{
Simulate Enrollment Time
}
